move Badge component
